jenkins docker lifecycle things
docker login will fail at this stage (local dev)

Since we don't keep our docker login in environment variable like this
on localhost, this will fail. If we were using a fully private registry
with password auth, and our upstream or syntax images were locally
hosted, then it would be important to first add this login back in here.

* add packages: cmake, pkg-config
* (fixup) can't interrupt escaped newlines with a comment
* CI scripts for testing and image tag inference

* abandon rails rake syntax stuff

This is just plain rake, we will include a plain old ruby module, there
is no autoloader so everything is explicit in here