an SSH deploy key has been written into jenkins ns
This key will get access to write in the "synths" branch (and other
branches) of our project repository. It is a secret that has been
deployed in our Jenkins namespace, using secrets manager or out-of-band
somehow. A generic secret with an SSH private key and known_hosts file.

Mounted on ~/.ssh