rebase spike-development-3
$GIT_REPO - variable for development purposes

it will be handy to be able to swap out this git repository easily from
here in the script, so let's make it a variable parameter in the script.

Was on the fence about running: cdk8s import k8s inside of CI,
opted to run it upstream intead.

npm run compile will succeed as long as k8s.ts is included
tests are used to ensure smoke-test readiness but are not comprehensive

* add package-lock.json - there is an error from CI about npm version

* Generally do not commit synths/*.js
(except for synths/jest.config.js, which is full of test configuration)

* do not include compilation targets
Generated by npm run compile and they should not be committed in CI

* do not save __snapshots__/*.test.ts.snap

I am not sure how snapshots work but they are recommended as a test
strategy in the distro files from cdk8s.  I do not know how jest works
outside of basic things, will use it to write some tests maybe, or skip
it. (It would be better to test this with something comprehensive before
it is applied to the cluster, but maybe kustomization's health checking
is good enough to avoid a complex suite of tests upstream...)

https://openconstructfoundation.org/testing-cdk8s/

So we can see a problem in our generated output thanks to test failures,
the namespace has not been mentioned in the re-hydrated yaml from cdk8s.

Let's add a test that asserts we include the namespace, and include it.

* run the tests in CI
this should stop synths commit from being generated in event that it
failed the test (this is imperfect, but it's a useful mitigation)

* do not use snapshots for now