enable github action
First, get nodejs from an action (don't make Jenkins responsible for
this anymore.)

Since we will use GitHub Actions, it's almost so simple that we can
just switch branch -- github actions has already done a clone, we can
throw the Dockerfile away.

We use `git checkout synths --` to disambiguate a path from the branch.

To the add-and-commit action, job step parameters go in a "with:" block

Todo (by GitHub action):
* update manifest
* show diff (if any)
* add to stage
* commit and push to `synths` branch